refactor(mobile): move the new-session prompt props to a types module
The base merge pushed new-session-prompt.tsx over the max-lines limit. The
props type is the only self-contained block, and it owned every type-only
import the component body never used.
